fix: std::max diff arg types template problem

+++ libmodbus_cpp/abstract_master.h
@@ -47,7 +47,7 @@ public:
 
 template<typename ValueType>
 ValueType AbstractMaster::readHoldingRegister(uint16_t address) {
-    int regCount = std::max(sizeof(ValueType) / sizeof(uint16_t), 1u);
+    int regCount = std::max(sizeof(ValueType) / sizeof(uint16_t), static_cast<size_t>(1u));
     ValueType result;
     int errorCode = modbus_read_registers(getBackend()->getCtx(), address, regCount, reinterpret_cast<uint16_t*>(&result));
     if (errorCode == -1)
@@ -57,7 +57,7 @@ ValueType AbstractMaster::readHoldingRegister(uint16_t address) {
 
 template<typename ValueType>
 void AbstractMaster::writeHoldingRegister(uint16_t address, ValueType value) {
-    int regCount = std::max(sizeof(ValueType) / sizeof(uint16_t), 1u);
+    int regCount = std::max(sizeof(ValueType) / sizeof(uint16_t), static_cast<size_t>(1u));
     int errorCode = modbus_write_registers(getBackend()->getCtx(), address, regCount, reinterpret_cast<uint16_t*>(&value));
     if (errorCode == -1)
         throw RemoteWriteError(modbus_strerror(errno));
@@ -65,7 +65,7 @@ void AbstractMaster::writeHoldingRegister(uint16_t address, ValueType value) {
 
 template<typename ValueType>
 ValueType AbstractMaster::readInputRegister(uint16_t address) {
-    int regCount = std::max(sizeof(ValueType) / sizeof(uint16_t), 1u);
+    int regCount = std::max(sizeof(ValueType) / sizeof(uint16_t), static_cast<size_t>(1u));
     ValueType result;
     int errorCode = modbus_read_input_registers(getBackend()->getCtx(), address, regCount, reinterpret_cast<uint16_t*>(&result));
     if (errorCode == -1)
